fix: changed the component logic to index and changed the order of the imports
