+++ src/components/CountryCard/index.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+interface CountryCardProps {}
+
+function CountryCard() {
+  return (
+    <div>
+      <div></div>
+      <div>
+        <p></p>
+        <p></p>
+        <p></p>
+        <p></p>
+      </div>
+    </div>
+  );
+}
+export default CountryCard;
